Updated to latest & moderns version, and used typescript
