--- .babelrc
@@ -1,5 +1,0 @@
-{
-  "presets": [
-    ["@babel/env", { "targets": { "node": 6 } }]
-  ]
-}
