+++ .travis.yml
@@ -1,5 +1,6 @@
 language: node_js
 node_js:
   - "node"
-  - "8"
-  - "6"
+  - "18"
+  - "16"
+  - "14"
