+++ jest.config.js
@@ -1,0 +1,5 @@
+/** @type {import('jest').Config} */
+const config = {
+};
+
+module.exports = config;
