+++ package.json
@@ -2,15 +2,15 @@
   "name": "rollup-plugin-uglify",
   "version": "6.0.4",
   "description": "Rollup plugin to minify generated bundle",
-  "main": "index.js",
-  "files": [
-    "index.js",
-    "transform.js"
-  ],
+  "main": "dist/index.js",
   "scripts": {
-    "test": "jest --runInBand",
-    "prepublish": "yarn test"
+    "test": "jest --config=./jest.config.js",
+    "prepublish": "yarn test",
+    "build": "rimraf ./dist && tsc"
   },
+  "files": [
+    "dist"
+  ],
   "repository": {
     "type": "git",
     "url": "git+https://github.com/TrySound/rollup-plugin-uglify.git"
@@ -26,15 +26,21 @@
   "dependencies": {
     "@babel/code-frame": "^7.18.6",
     "jest-worker": "^29.5.0",
+    "rimraf": "^4.4.0",
     "serialize-javascript": "^6.0.1",
+    "typescript": "^5.0.2",
     "uglify-js": "^3.17.4"
   },
   "peerDependencies": {
-    "rollup": ">=2"
+    "rollup": ">=3"
   },
   "devDependencies": {
     "@babel/core": "^7.21.3",
     "@babel/preset-env": "^7.20.2",
+    "@babel/preset-typescript": "^7.21.0",
+    "@types/babel-code-frame": "^6.20.5",
+    "@types/serialize-javascript": "^5.0.2",
+    "@types/uglify-js": "^3.17.1",
     "babel-jest": "^29.5.0",
     "jest": "^29.5.0",
     "prettier": "^2.8.5",
