+++ src/index.ts
@@ -1,0 +1,2 @@
+export {uglify} from './uglify';
+export {transform} from './transform';
