--- test/fixtures/sourcemap.js
@@ -1,3 +1,0 @@
-import result from './export-number.js';
-
-console.log(result);
